chore: update to new file structure
